Copy in seq2seq style attentive GRU encoder + decoder code.
Primarily sourced from:

	https://bastings.github.io/annotated_encoder_decoder/

This seems to be almost exactly what we need. It differs in some small
ways from the GRU4REC paper (attention function being a glaring
example), but otherwise seems very well adapted to what we're trying to
do here.

Note that this is a sequence-to-sequence model! The *Attention is All
you Need* paper is great, but uses feedforward nets only (it gets the
time dependence via some positional encoding magic, which we're not
doing here) -- so we'll explore attention via sequence-to-sequence
stuff, then slap it onto our BERT4REC model above.